# Conflicts:
#	src/Client/Views/AddBookingView/VehicleViewCell/VehicleViewCell.fxml

+++ src/Client/Model/Booking.java
@@ -3,15 +3,17 @@ package Client.Model;
 
 import java.util.GregorianCalendar;
 
+import java.util.GregorianCalendar;
+
 public class Booking
 {
   private int idOfCustomer;
   private String licencePlate;
   private GregorianCalendar startTime;
   private GregorianCalendar endTime;
-  private int price;
+  private double price;
 
-  public Booking(int idOfCustomer,String licencePlate,GregorianCalendar startTime,GregorianCalendar endTime,int price){
+  public Booking(int idOfCustomer,String licencePlate,GregorianCalendar startTime,GregorianCalendar endTime,double price){
     this.idOfCustomer = idOfCustomer;
     this.licencePlate = licencePlate;
     this.startTime = startTime;
@@ -19,12 +21,12 @@ public class Booking
     this.price = price;
   }
 
-  public void setPrice(int price)
+  public void setPrice(double price)
   {
     this.price = price;
   }
 
-  public int getPrice()
+  public double getPrice()
   {
     return price;
   }
